created abstract class
