+++ app/src/main/java/com/agrawalsuneet/loaders/MainActivity.kt
@@ -4,12 +4,14 @@ import android.os.Bundle
 import android.os.Handler
 import android.support.v4.content.ContextCompat
 import android.support.v7.app.AppCompatActivity
+import android.view.View
 import android.view.animation.Animation
 import android.view.animation.BounceInterpolator
 import android.view.animation.LinearInterpolator
 import android.view.animation.RotateAnimation
 import android.widget.LinearLayout
 import com.agrawalsuneet.loaderspack.loaders.*
+import kotlinx.android.synthetic.main.activity_main_multipleripple.*
 
 class MainActivity : AppCompatActivity() {
 
@@ -17,7 +19,7 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.main_search)
+        setContentView(R.layout.activity_main_multipleripple)
 
         supportActionBar?.setTitle("SearchLoader")
 
@@ -39,6 +41,12 @@ class MainActivity : AppCompatActivity() {
 
         //initGaugeLoader()
         //initSearchLoader()
+
+        btn.setOnClickListener(object : View.OnClickListener{
+            override fun onClick(p0: View?) {
+                multipleripple.visibility = if( multipleripple.visibility == View.VISIBLE)  View.INVISIBLE else View.VISIBLE
+            }
+        })
     }
 
     private fun initSearchLoader() {
