+++ app/src/main/res/layout/activity_main_multipleripple.xml
@@ -7,7 +7,7 @@
     android:gravity="center"
     android:orientation="vertical">
 
-    <LinearLayout
+    <!--<LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:gravity="center"
@@ -105,6 +105,29 @@
             android:scaleType="centerCrop"
             android:src="@mipmap/ic_launcher_round" />
 
-    </RelativeLayout>
+    </RelativeLayout>-->
+
+    <com.agrawalsuneet.loaderspack.loaders.MultipleRippleLoader
+        android:id="@+id/multipleripple"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        app:multipleripple_animDuration="3000"
+        android:layout_margin="16dp"
+        app:multipleripple_circleColor="@color/blue_selected"
+        app:multipleripple_circleInitialRadius="40dp"
+        app:multipleripple_fromAlpha="0.9"
+        app:multipleripple_interpolator="@android:anim/linear_interpolator"
+        app:multipleripple_noOfRipples="3"
+        android:visibility="invisible"
+        app:multipleripple_toAplha="0.0" />
+
+    <Button
+        android:id="@+id/btn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:padding="16dp"
+        android:text="Animate"
+        android:layout_margin="16dp"/>
 
 </LinearLayout>
